requirements.txt
opencv_python==4.1.1.26
numpy==1.17.3
tensorflow==2.0.1
imgaug==0.3.0
pandas==0.25.2
fastapi==0.49.0
PyYAML==5.3
scikit_learn==0.22.2
